refactor(examples): refactor example directory
