+++ examples/document.mock.json
@@ -2,6 +2,7 @@
 	"id": "XvoFFREAAM0WGBng",
 	"uid": "test",
 	"type": "page",
+	"url": null,
 	"href": "https://example.cdn.prismic.io/api/v2/documents/search",
 	"tags": [],
 	"first_publication_date": "2020-06-29T15:13:27+0000",
